fix(stella-store): move the scope backfill into its own module
The file-size gate flagged usage.rs growing 94 lines past its
grandfathered ceiling. The backfill (#406) now lives in usage/backfill.rs
with its witness test; usage.rs grows only by the module declaration,
recorded in the baseline as the gate's message directs for irreducible
growth.

Refs #406
