+++ scripts/file-size-baseline.txt
@@ -24,7 +24,7 @@
 2399 stella-protocol/src/event.rs
 2246 stella-store/src/lib.rs
 2200 stella-store/src/tests.rs
-1890 stella-store/src/usage.rs
+1893 stella-store/src/usage.rs
 1556 stella-tools/src/media.rs
 3054 stella-tools/src/registry.rs
 1829 stella-tools/src/scripts.rs
