+++ stella-store/src/usage.rs
@@ -510,50 +510,6 @@ impl UsageStore {
         Ok(out)
     }
 
-    /// Re-stamp this project's NULL-org hub rows with the just-registered
-    /// scope, in one transaction — the backfill `stella cloud register` runs
-    /// so history replicated *before* registration joins the org's reporting
-    /// and drain instead of staying invisible forever (#406).
-    ///
-    /// Policy for rows that were in the hub under NULL org: they were never
-    /// shipped — [`Self::cloud_pending`] is org-scoped and skips NULL-org
-    /// rows — so re-stamping cannot double-send anything. The rows are also
-    /// *moved* to fresh rowids above the table's current max: the org's drain
-    /// cursor only ever advances (`rowid > cursor`), and an org that already
-    /// acked rows from other workspaces would otherwise never see backfilled
-    /// history parked below its cursor. Rowids carry no meaning beyond drain
-    /// order (the stable key is `(project_id, source_rowid)`), and no cursor
-    /// or quarantine entry can reference a NULL-org row.
-    ///
-    /// Idempotent — a second run finds no NULL-org rows for the project and
-    /// changes nothing — and scoped to one `project_id`; another workspace's
-    /// rows are never touched. A row with a non-NULL `workspace_id` keeps it.
-    ///
-    /// Returns how many rows were re-stamped.
-    pub fn backfill_scope(
-        &self,
-        project_id: &str,
-        org_id: &str,
-        workspace_id: Option<&str>,
-    ) -> Result<u64> {
-        let conn = self.lock();
-        let tx = conn.unchecked_transaction()?;
-        let max_rowid: i64 =
-            tx.query_row("SELECT COALESCE(MAX(rowid), 0) FROM telemetry", [], |r| {
-                r.get(0)
-            })?;
-        let changed = tx.execute(
-            "UPDATE telemetry \
-                SET rowid = rowid + ?1, \
-                    org_id = ?2, \
-                    workspace_id = COALESCE(workspace_id, ?3) \
-              WHERE project_id = ?4 AND org_id IS NULL",
-            params![max_rowid, org_id, workspace_id, project_id],
-        )?;
-        tx.commit()?;
-        Ok(changed as u64)
-    }
-
     /// Hub rows for one org not yet acknowledged by the cloud, oldest first
     /// — the drain a cloud syncer walks before [`Self::ack_cloud_synced`].
     pub fn cloud_pending(&self, org_id: &str, limit: usize) -> Result<Vec<CloudTelemetryEvent>> {
@@ -1178,6 +1134,9 @@ pub struct QuarantinedRow {
     pub quarantined_at: String,
 }
 
+// The registration-time scope backfill (#406), in its own module.
+mod backfill;
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -1362,56 +1321,6 @@ mod tests {
             hub.cloud_pending("acme", 10).unwrap().is_empty(),
             "an out-of-order ack never rewinds the cursor"
         );
-    }
-
-    /// Witness for #406: rows replicated before `stella cloud register` land
-    /// NULL-org and used to stay NULL forever — invisible to org reporting
-    /// and permanently excluded from the cloud drain.
-    #[test]
-    fn backfill_restamps_pre_registration_rows_into_org_scope() {
-        let hub = UsageStore::in_memory().unwrap();
-        // History replicated while unregistered → NULL org.
-        hub.replicate_telemetry(
-            &scope(None, None),
-            &[source_row(1, 0.01), source_row(2, 0.02)],
-        )
-        .unwrap();
-        // Another org workspace has already drained and acked rows, so the
-        // org cursor sits past the historical rows' rowids.
-        let mut other = scope(Some("acme"), Some("ws-other"));
-        other.project_id = "proj_b".into();
-        hub.replicate_telemetry(&other, &[source_row(1, 0.05)])
-            .unwrap();
-        let acked = hub.cloud_pending("acme", 10).unwrap();
-        hub.ack_cloud_synced("acme", acked.last().unwrap().hub_rowid)
-            .unwrap();
-
-        // Register proj_a into acme: its history joins the org report…
-        assert_eq!(
-            hub.backfill_scope("proj_a", "acme", Some("ws-1")).unwrap(),
-            2
-        );
-        let totals = hub.global_telemetry_totals(Some("acme")).unwrap();
-        assert_eq!(totals[0].calls, 3, "historical rows join the org report");
-
-        // …and the drain, even though the org cursor already advanced past
-        // the rowids the historical rows used to sit at.
-        let pending = hub.cloud_pending("acme", 10).unwrap();
-        assert_eq!(pending.len(), 2, "backfilled rows become drainable");
-        assert!(pending.iter().all(|p| p.project_id == "proj_a"));
-        assert!(
-            pending
-                .iter()
-                .all(|p| p.workspace_id.as_deref() == Some("ws-1")),
-            "NULL workspace_id is stamped alongside org_id"
-        );
-
-        // Idempotent: nothing left to re-stamp, nothing double-processed.
-        assert_eq!(
-            hub.backfill_scope("proj_a", "acme", Some("ws-1")).unwrap(),
-            0
-        );
-        assert_eq!(hub.cloud_pending("acme", 10).unwrap().len(), 2);
     }
 
     // ---- poison-row quarantine (#467) --------------------------------------
